Minimize files to change to support new service
